requirements.txt
langchain
langchain-community
telegram
python-telegram-bot
chromadb==0.3.29
google-generativeai 
langchain-google-genai 
python-dotenv 
pypdf
flask
